op-e2e: Use helpers package under proofs rather than a completely separate package tree for helpers.
