+++ op-e2e/actions/proofs/helpers/matrix.go
@@ -1,4 +1,4 @@
-package proofs
+package helpers
 
 import (
 	"fmt"
